telegramBot moved from model to its own telegramBot package
