--- src/main/java/pro/sky/javacourse/AnimalShelterBot/model/menu/BotMenuBuilder.java
@@ -1,9 +1,0 @@
-package pro.sky.javacourse.AnimalShelterBot.model.menu;
-
-public class BotMenuBuilder {
-
-
-
-
-
-}
