+++ src/main/java/pro/sky/javacourse/AnimalShelterBot/telegramBot/menu/BotMenuBuilder.java
@@ -1,0 +1,9 @@
+package pro.sky.javacourse.AnimalShelterBot.telegramBot.menu;
+
+public class BotMenuBuilder {
+
+
+
+
+
+}
